Database fixes

See merge request CommunityOutreach/winter-2020/philippefullsack_alliancefrancaise!23

+++ app/src/main/java/com/csci/afevents/entities/Event.java
@@ -4,9 +4,9 @@ public class Event {
     private String eventId;
     private String eventName;
     private String description;
-    private long date;
+    private int date;
 
-    public Event(String eventId, String eventName, String description, long date) {
+    public Event(String eventId, String eventName, String description, int date) {
         this.eventId = eventId;
         this.eventName = eventName;
         this.description = description;
@@ -21,7 +21,7 @@ public class Event {
     public String getDescription(){
         return description;
     }
-    public long getDate(){
+    public int getDate(){
         return date;
     }
 }
